requirements.txt
pyqt5
PyGithub
crumbs
tqdm